Reorganization, get sprite pixel method
